backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
bcrypt==4.1.2
PyJWT==2.8.0
redis==5.0.1
celery==5.3.4
python-magic==0.4.27